+++ python/snips_nlu_parsers/gazetteer_entity_parser.py
@@ -63,7 +63,7 @@ class GazetteerEntityParser(object):
                                    "gazetteer entity parser")
         return cls(parser)
 
-    def parse(self, text, scope=None):
+    def parse(self, text, scope=None, max_alternative_resolved_values=5):
         """Extract gazetteer entities from *text*
 
         Args:
@@ -72,6 +72,9 @@ class GazetteerEntityParser(object):
                 the parser will extract entities using the provided scope
                 instead of the entire scope of all available entities. This
                 allows to look for specifics entities.
+            max_alternative_resolved_values (int, optional): Maximum number of
+                alternative resolved values to return in addition to the top
+                one (default 5).
 
         Returns:
             list of dict: The list of extracted entities
@@ -91,7 +94,8 @@ class GazetteerEntityParser(object):
 
         with string_pointer(c_char_p()) as ptr:
             exit_code = lib.snips_nlu_parsers_extract_gazetteer_entities_json(
-                self._parser, text.encode("utf8"), scope, byref(ptr))
+                self._parser, text.encode("utf8"), scope,
+                max_alternative_resolved_values, byref(ptr))
             check_ffi_error(exit_code, "Something went wrong when "
                                        "extracting gazetteer entities")
             result = string_at(ptr)